Prevent docker run command from including arbitrary bash commands or redirection
